requirements.txt
dotenv==0.9.9
openai==1.77.0
playwright==1.52.0
azure-identity==1.22.0
